CombatCard moved to folder container
